# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateShortLinkRequest:
    properties:
      originalUrl:
        type: string
    required:
    - originalUrl
    type: object
  models.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  models.UpdateShortLinkRequest:
    properties:
      isActive:
        type: boolean
      originalUrl:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      fullName:
        type: string
      id:
        type: integer
      profilePhoto:
        type: string
    type: object
  response.ResponseError:
    properties:
      error:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.ResponseSuccess:
    properties:
      data: {}
      message:
        example: Success message
        type: string
      success:
        example: true
        type: boolean
    type: object
  services.DashboardStats:
    properties:
      avgClickRate:
        format: float64
        type: number
      last7DaysStat: {}
      totalLinks:
        type: integer
      totalVisits:
        type: integer
    type: object
info:
  contact: {}
  description: Dokumentasi REST API menggunakan Gin dan Swagger
  title: API Koda Shortlink Documentation
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Log in with existing email data
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        format: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate refresh token from httpOnly cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Get new access token using refresh token from httpOnly cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseSuccess'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new user with a unique email
      parameters:
      - description: Full name user
        in: formData
        name: fullname
        required: true
        type: string
      - description: Email user
        in: formData
        name: email
        required: true
        type: string
      - description: Password user
        format: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      summary: Register new user
      tags:
      - auth
  /dashboard/stats:
    get:
      consumes:
      - application/json
      description: Retrieve user-specific statistics for dashboard overview
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/services.DashboardStats'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Get dashboard statistics
      tags:
      - dashboard
  /links:
    get:
      consumes:
      - application/json
      description: Get all short links created by authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Get all short links
      tags:
      - links
    post:
      consumes:
      - application/json
      description: Create a new short link with auto-generated code
      parameters:
      - description: Short link details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateShortLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Create short link
      tags:
      - links
  /links/{shortCode}:
    delete:
      consumes:
      - application/json
      description: Delete short link by short code
      parameters:
      - description: Short code
        in: path
        name: shortCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Delete short link
      tags:
      - links
    get:
      consumes:
      - application/json
      description: Get specific short link details by short code
      parameters:
      - description: Short code
        in: path
        name: shortCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Get short link by code
      tags:
      - links
    put:
      consumes:
      - application/json
      description: Update short link details (original URL and/or active status)
      parameters:
      - description: Short code
        in: path
        name: shortCode
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateShortLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Update short link
      tags:
      - links
  /users:
    get:
      description: Get specific user detail by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseError'
      security:
      - BearerAuth: []
      summary: Get user detail
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"